Create Lab_Online_on_Linked_List
Question is:
You will be given two linked list. You will have to traverse the lists sequentially as the sum of each node of linked list 1 plus linked list 2 should be in ascending mode so that you can take the value as valid for your new linked list either not. suppose the sum of linked list 1 and linked list 2 for first node is 3 but in case of second node  the sum is 1 the you can't take the value of second node if the value is equal of greater then 3 then you can take it as a valid input for your new linked list. And the number of  total given values may different of given lists.